+++ __tests__/mask/money.mask.test.js
@@ -113,16 +113,6 @@ test('1 suffixUnit $$$ results R$0,01', () => {
     expect(received).toBe(expected)
 })
 
-test('1 zeroCents results R$1,00', () => {
-    var mask = new MoneyMask()
-    var expected = 'R$1,00'
-    var received = mask.getValue('1', {
-        zeroCents: true
-    })
-
-    expect(received).toBe(expected)
-})
-
 test('US$ config with value 1234567 results US$12,345.67', () => {
     var mask = new MoneyMask()
     var expected = 'US$12,345.67'
@@ -153,20 +143,6 @@ test('111111 results R$1.111,11 and raw value 1111.11', () => {
     var received = mask.getValue('111111')
 
     var expectedRawValue = 1111.11
-    var receivedRawValue = mask.getRawValue(received)
-
-    expect(received).toBe(expected)
-    expect(receivedRawValue).toBe(expectedRawValue)
-})
-
-test('1 zeroCents results R$1,00 and raw value 1', () => {
-    var mask = new MoneyMask()
-    var expected = 'R$1,00'
-    var received = mask.getValue('1', {
-        zeroCents: true
-    })
-
-    var expectedRawValue = 1
     var receivedRawValue = mask.getRawValue(received)
 
     expect(received).toBe(expected)